# perf: implement custom value to reduce memory allocation