… bitwise_test.go
… go.sum
… visitor.go
… visitor_test.go